Simplify settings module
