+++ presentation/app/__init__.py
@@ -1,0 +1,3 @@
+"""Application package for HobbyPicker."""
+from .main import start_app
+__all__ = ["start_app"]
