+++ presentation/app/ui_components/__init__.py
@@ -1,0 +1,1 @@
+"""UI component builders used by the application."""
